Show message to user when logging in.
It takes a few seconds for the login process to redirect the
user to the logged in page.
This can be confusing to the user who might think there's something
wrong with the login process.
Also, the user is not supposed to click login again while this happens,
so the login buttons are hidden.